requirements.txt

# LangChain Core
langchain>=0.1.0
langchain-ollama>=0.1.0
langchain-community>=0.1.0

# Utilities
python-dotenv>=1.0.0

# Search Tools
duckduckgo-search>=4.0.0
wikipedia>=1.4.0